+++ docs/changelog.rst
@@ -21,6 +21,7 @@ in specific versions.
     - Example: ``queue.sort(key=lambda track: track.title)`` to sort by track title.
   - :meth:`Queue.reverse` to reverse the order of the queue.
   - :meth:`Queue.dedupe` to remove duplicate tracks from the queue.
+  - :meth:`Queue.__reversed__` to allow the built-in :func:`reversed` function on the queue.
 - Added :meth:`Player.skip_to` to skip directly to a track at a specific
   queue index.
 - Added :meth:`Queue.move` to move a track from one index to another, shifting